+++ src/constants/index.ts
@@ -5,3 +5,4 @@ export * from "./statuses";
 export * from "./phases";
 export * from "./game";
 export * from "./avatar-scripts";
+export * from "./host";